Lots of refactoring and movement Adding plugin to help with consistent alias
