--- app/components/useMountedState.tsx
@@ -1,16 +1,0 @@
-import { useEffect, useCallback, useRef } from "react";
-
-export function useMountedState(): () => boolean {
-  const mountedRef = useRef<boolean>(false);
-  const get = useCallback(() => mountedRef.current, []);
-
-  useEffect(() => {
-    mountedRef.current = true;
-
-    return () => {
-      mountedRef.current = false;
-    };
-  }, []);
-
-  return get;
-}
