+++ app/entry.client.tsx
@@ -5,8 +5,7 @@
  */
 
 import { RemixBrowser } from "@remix-run/react";
-
-import { StrictMode, startTransition } from "react";
+import { startTransition, StrictMode } from "react";
 import { hydrateRoot } from "react-dom/client";
 
 // load service worker @ sw.js
