+++ app/features/comments/HnComment.tsx
@@ -1,13 +1,15 @@
-import { cn, isNavigator, timeSince } from "@/utils";
 import { decode } from "html-entities";
 import { ArrowUpRightFromSquare } from "lucide-react";
 import { useEffect, useRef, useState } from "react";
 import sanitizeHtml from "sanitize-html";
-import { useCommentStore } from "~/features/comments/indexedDb";
+
+import { cn, isNavigator, timeSince } from "@/utils";
+import { StoryContext } from "~/features/storyPage/HnStoryPage";
+import { isValidComment } from "~/lib/isValidComment";
+import { useCommentStore } from "~/stores/useCommentStore";
 import { KidsObj3, useDataStore } from "~/stores/useDataStore";
+
 import { HnCommentList } from "./HnCommentList";
-import { StoryContext } from "./HnStoryPage";
-import { isValidComment } from "./isValidComment";
 
 export interface HnCommentProps {
   comment: KidsObj3 | null;
