+++ app/features/comments/HnCommentList.tsx
@@ -1,4 +1,5 @@
 import { KidsObj3 } from "@/stores/useDataStore";
+
 import { HnComment } from "./HnComment";
 import { InfiniteScrollContainer } from "./InfiniteScrollContainer";
