+++ app/features/storyList/HnStoryList.tsx
@@ -1,7 +1,10 @@
-import { HnStorySummary, StoryPage, useDataStore } from "@/stores/useDataStore";
-import { HnListItem } from "./HnListItem";
-import { useSortFunction } from "./useSortFunction";
 import { useEffect } from "react";
+
+import { HnStorySummary, StoryPage, useDataStore } from "@/stores/useDataStore";
+import { useSortFunction } from "~/hooks/useSortFunction";
+
+import { HnListItem } from "./HnListItem";
+
 
 interface HnStoryListProps {
   items?: HnStorySummary[];
