+++ app/lib/getCleanPathName.tsx
