+++ app/lib/getColorsForStory.tsx
@@ -1,4 +1,5 @@
 import { orderBy, uniq } from "lodash-es";
+
 import { HnItem, KidsObj3 } from "~/stores/useDataStore";
 
 export function getColorsForStory(story: HnItem): Record<string, string> {
