+++ app/lib/processHtmlAndTruncateAnchorText.tsx
