+++ app/routes/api.topstories.$type.tsx
@@ -1,4 +1,5 @@
-import { LoaderFunctionArgs, json } from "@remix-run/node";
+import { json, LoaderFunctionArgs } from "@remix-run/node";
+
 import { STORY_TYPE } from "~/models/interfaces";
 import { cachedData } from "~/server/server";
