+++ app/routes/useGetLocalData.tsx
@@ -1,4 +1,5 @@
 import { useEffect, useState } from "react";
+
 import { HnStorySummary, useDataStore } from "~/stores/useDataStore";
 
 export function useGetLocalData() {
